+++ crates/fs_server/src/fuse_server.rs
@@ -1,6 +1,5 @@
 use fractal_fuse::*;
 use std::ffi::OsStr;
-use std::os::fd::OwnedFd;
 use std::sync::Arc;
 
 use crate::error::FsError;
@@ -38,8 +37,8 @@ fn fs_err(e: FsError) -> Errno {
 }
 
 impl Filesystem for FuseServer {
-    async fn init(&self, _req: Request, fuse_dev_fd: Arc<OwnedFd>) -> FsResult<ReplyInit> {
-        self.vfs.vfs_init(fuse_dev_fd);
+    async fn init(&self, _req: Request) -> FsResult<ReplyInit> {
+        self.vfs.vfs_init();
         Ok(ReplyInit {
             max_write: 1024 * 1024,
             ..Default::default()